+++ src/pydbm/__init__.py
@@ -1,7 +1,9 @@
-from pydbm.exceptions import DoesNotExists, OdbmBaseException, OdbmTypeError
-from pydbm.models.base import BaseModel
-from pydbm.models.fields import Field
-from pydbm.models.validators import (
+from pydbm.exceptions import BaseException, DoesNotExists, PydbmTypeError, ValidationError
+from pydbm.models import (
+    AutoField,
+    BaseModel,
+    Field,
+    Undefined,
     validate_bool,
     validate_bytes,
     validate_date,
@@ -19,11 +21,14 @@ from pydbm.models.validators import (
 )
 
 __all__ = (
-    "BaseModel",
     "DoesNotExists",
+    "BaseException",
+    "PydbmTypeError",
+    "ValidationError",
+    "BaseModel",
+    "AutoField",
     "Field",
-    "OdbmBaseException",
-    "OdbmTypeError",
+    "Undefined",
     "validate_bool",
     "validate_bytes",
     "validate_date",